requirements.txt
asgiref==1.1.2
astroid==2.5.6
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
cffi==1.14.5
channels==1.1.8
constantly==15.1.0
cryptography==3.4.7
daphne==1.4.2
Django==2.2.20
django-rest-framework==0.1.0
djangorestframework==3.12.4
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pathlib==1.0.1
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pytz==2021.1
semantic-version==2.8.5
service-identity==18.1.0
setuptools-rust==0.12.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
wrapt==1.12.1
zope.interface==5.4.0
